requirements.txt
python-dotenv==0.19.2
tweepy==4.8.0